requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.4
defusedxml==0.7.0rc2
distlib==0.3.0
Django==3.0.5
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
peewee==3.14.1
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.4.4
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.14.0
social-auth-core==4.0.3
-e git://github.com/dmlb2000/social-storage-peewee.git@b0654a38937bfbd98cc16879831f6153b513678e#egg=social_auth_storage_peewee
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.3
virtualenv==20.0.18
wrapt==1.11.2
zipp==3.1.0
